adapt init_logging docstring to numpy format
